polymarket_copiloto_pro_v2/requirements.txt
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
rich>=13.7.0
